refactor(fe): 회사 등록 빈값 null 처리 수정 및 이미지들 assets 폴더로 이동
