+++ frontend/src/routes/login/index.tsx
@@ -3,6 +3,7 @@ import { useEffect, useState } from 'react';
 import { createFileRoute, useRouter } from '@tanstack/react-router';
 import { ChevronLeft } from 'lucide-react';
 
+import landingImage from '@/assets/images/landing-1.jpg';
 import ErrorMessage from '@/components/auth/ErrorMessage';
 import LoginForm from '@/components/auth/LoginForm';
 import SocialLoginButtons from '@/components/auth/SocialLoginButtons';
@@ -81,7 +82,7 @@ function LoginPage() {
   return (
     <div className="min-h-screen flex">
       <div className="hidden lg:block lg:w-2/3">
-        <img src="/images/landing-1.jpg" alt="login-background" className="w-full h-full" />
+        <img src={landingImage} alt="login-background" className="w-full h-full" />
       </div>
 
       <div className="w-full lg:w-1/2 flex items-center justify-center p-8 relative">
